requirements.txt
pyyaml>=6.0
pyserial>=3.5
numpy>=1.24
requests>=2.31
